Move analysis functions into engine.cpp
file_analyze, establish_argument_boundaries, file_refactor,
garbage_collect and their helpers are analysis logic that belongs
neither to the web interface nor to startup options. Moving them
to engine.cpp is the first step toward a proper CscoutEngine class.

Model: claude-sonnet-4-6
Prompt: I need engine.h updated to declare RefFunCall, EdgeMatrix,
extern counters, and free function declarations so cscout.cpp can
see symbols defined in engine.cpp.
Output: Provided the updated engine.h with all required declarations.
